oneshotiterator class
