+++ tensorflow-frameworks/tensorflow-data/src/test/java/org/tensorflow/data/DatasetTest.java
@@ -11,6 +11,9 @@ import java.util.List;
 import static org.junit.Assert.assertArrayEquals;
 
 public class DatasetTest extends DatasetTestBase {
+  private static void main(String[] args) {
+    new DatasetTest().testEagerDatasetIterator();
+  }
 
   @Test
   public void testEagerDatasetIterator() {
@@ -45,9 +48,9 @@ public class DatasetTest extends DatasetTestBase {
                   tf.constant(testMatrix2)),
               Arrays.asList(TInt32.DTYPE, TInt32.DTYPE));
 
-      Pair<Operation, List<Output<?>>> graphIteratorComponents = dataset.makeOneShotIterator();
-      Operation makeIterator = graphIteratorComponents.first();
-      List<Output<?>> components = graphIteratorComponents.second();
+      OneShotIterator oneShotIterator = dataset.makeOneShotIterator();
+      Operation makeIterator = oneShotIterator.getMakeIteratorOp();
+      List<Output<?>> components = oneShotIterator.getComponents();
 
       try (Session session = new Session(graph)) {
         session.runner()
@@ -76,6 +79,4 @@ public class DatasetTest extends DatasetTestBase {
       }
     }
   }
-
-
 }
